+++ src/main/java/com/todaysound/todaysound_server/global/dto/PageRequest.java
@@ -1,6 +1,6 @@
 package com.todaysound.todaysound_server.global.dto;
 
-public record PageRequest(Long page, Integer size) {
+public record PageRequest(Integer page, Integer size) {
     private static final Integer DEFAULT_SIZE = 5;
 
     @Override
